+++ src/AudioLoading/IAudioClipLoader.cs
@@ -1,7 +1,0 @@
-using UnityEngine;
-using System.Threading.Tasks;
-
-public interface IAudioClipLoader
-{
-    public Task<AudioClip> LoadAudioClipAsync(string path);
-}